Use floats in steering and throttle control fields
Prior to this commit, the fields of the module's control state
structure were using data types that matches what would eventually be
sent out in the report, but this lost some precision that made the PID
controller operate more smoothly. This commit uses floats for those
fields and casts them to the appropriate data types during report
publishing.

+++ platforms/kia_soul/firmware/brake/include/brake_control.h
@@ -153,10 +153,10 @@ typedef struct
     float current_sensor_brake_pressure; /* Current brake pressure as read
                                                from the brake pressure sensor. */
 
-    int16_t current_vehicle_brake_pressure; /* Current brake pressure as reported
+    float current_vehicle_brake_pressure; /* Current brake pressure as reported
                                                by the vehicle. */
 
-    uint16_t commanded_pedal_position; /* Brake pedal position commanded by
+    float commanded_pedal_position; /* Brake pedal position commanded by
                                           controller. */
 } kia_soul_brake_control_state_s;
